Refactoring: Separate service interfaces and implementations directory location
